% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nhanes_qc_result.R
\name{print.nhanes_qc_result}
\alias{print.nhanes_qc_result}
\title{Print method for nhanes_qc_result objects}
\usage{
\method{print}{nhanes_qc_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{nhanes_qc_result}, as returned by \code{\link[=nhanes_qc]{nhanes_qc()}}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for nhanes_qc_result objects
}
